refactor: Move useDashboard logic hooks into hooks/dashboard and Update import statements
